requirements.txt
py-clob-client>=0.13.0
websockets>=12.0
scipy>=1.11.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
